Rename controllers to view-models
