+++ task_2_3_1/src/main/java/kozlov/kirill/snake/view_model/GameOverViewModel.java
@@ -1,8 +1,7 @@
-package kozlov.kirill.snake.model_view;
+package kozlov.kirill.snake.view_model;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
-import javafx.scene.text.Text;
 import kozlov.kirill.snake.ExcludeClassFromJacocoGeneratedReport;
 import kozlov.kirill.snake.model.Model;
 import kozlov.kirill.snake.model.game.GameModel;
@@ -10,7 +9,7 @@ import kozlov.kirill.snake.view.SceneEnum;
 import kozlov.kirill.snake.view.SceneManager;
 
 @ExcludeClassFromJacocoGeneratedReport
-public class GameOverController implements SceneManagerAccessible {
+public class GameOverViewModel implements SceneManagerAccessible {
     private SceneManager sceneManager;
 
     @FXML
