+++ task_2_3_1/src/main/java/kozlov/kirill/snake/view_model/SceneManagerAccessible.java
@@ -1,4 +1,4 @@
-package kozlov.kirill.snake.model_view;
+package kozlov.kirill.snake.view_model;
 
 import kozlov.kirill.snake.view.SceneManager;
