+++ task_2_3_1/src/main/java/kozlov/kirill/snake/view_model/settings/Validator.java
@@ -1,4 +1,4 @@
-package kozlov.kirill.snake.model_view.settings;
+package kozlov.kirill.snake.view_model.settings;
 
 import javafx.scene.control.TextInputControl;
 import javafx.scene.input.KeyEvent;
