+++ task_2_3_1/src/main/resources/kozlov/kirill/snake/game-scene.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.control.Button?>
 
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="kozlov.kirill.snake.model_view.GameController">
+      fx:controller="kozlov.kirill.snake.view_model.GameViewModel">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
