+++ task_2_3_1/src/main/resources/kozlov/kirill/snake/menu-scene.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.control.Button?>
 
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="kozlov.kirill.snake.model_view.MenuController">
+      fx:controller="kozlov.kirill.snake.view_model.MenuViewModel">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
