+++ task_2_3_1/src/main/resources/kozlov/kirill/snake/settings-scene.fxml
@@ -10,7 +10,7 @@
 <?import javafx.geometry.Insets?>
 <VBox alignment="CENTER" spacing="20.0"
       xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="kozlov.kirill.snake.model_view.settings.SettingsController">
+      fx:controller="kozlov.kirill.snake.view_model.settings.SettingsViewModel">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
